requirements.txt
django==3.1
djangorestframework==3.11.1
PyMySQL==0.10.0
django-bootstrap3==14.1.0
bs4
pymorphy2
requests
configobj
lxml